refactor(native): centralize project setup logic into utils.js (DRY)
