+++ src/native/androidSetup.js
@@ -1,7 +1,7 @@
 import fs from "fs"
 import path from "path"
 import { exec } from "child_process"
-import { runCommand, validateAndCompleteConfig, promptUserWithTimeout } from "./utils.js"
+import { runCommand, validateAndCompleteConfig, initializeConfig, saveConfig, handlePostSetupSteps } from "./utils.js"
 import TerminalProgress from "./TerminalProgress.js"
 import { getLocalIPAddress, updateConfigWithLocalIP, isServerRunning, startServerBackground } from "./setupServer.js"
 
@@ -140,43 +140,6 @@ async function validateJavaEnvironment() {
     progress.complete("java")
 }
 
-async function initializeConfig() {
-    const configFile = fs.readFileSync(configPath, "utf8")
-    const config = JSON.parse(configFile)
-    const { WEBVIEW_CONFIG } = config
-
-    if (!WEBVIEW_CONFIG || Object.keys(WEBVIEW_CONFIG).length === 0) {
-        progress.log("WebView Config missing in " + configPath, "error")
-        process.exit(1)
-    }
-
-    if (!WEBVIEW_CONFIG.android) {
-        WEBVIEW_CONFIG.android = {}
-    }
-
-    return { WEBVIEW_CONFIG }
-}
-
-async function saveConfig(newConfig) {
-    try {
-        const existingConfigFile = fs.readFileSync(configPath, "utf8")
-        const existingConfig = JSON.parse(existingConfigFile)
-
-        const updatedConfig = {
-            ...existingConfig,
-            WEBVIEW_CONFIG: {
-                ...existingConfig.WEBVIEW_CONFIG,
-                android: newConfig.WEBVIEW_CONFIG.android,
-            },
-        }
-
-        fs.writeFileSync(configPath, JSON.stringify(updatedConfig, null, 2))
-        progress.log("Configuration saved successfully", "success")
-    } catch (error) {
-        progress.log("Failed to save configuration: " + error.message, "error")
-        process.exit(1)
-    }
-}
 
 function validateAndroidTools(androidConfig) {
     const ANDROID_SDK = androidConfig.sdkPath
@@ -264,7 +227,7 @@ async function setupAndroidEnvironment() {
         await validateJavaEnvironment()
 
         progress.start("config")
-        await initializeConfig()
+        await initializeConfig("android", configPath, progress)
         const config = await validateAndCompleteConfig("android", configPath)
         progress.complete("config")
 
@@ -287,55 +250,15 @@ async function setupAndroidEnvironment() {
         }
 
         progress.start("saveConfig")
-        await saveConfig({ WEBVIEW_CONFIG: config })
+        await saveConfig("android", configPath, { WEBVIEW_CONFIG: config }, progress)
         progress.complete("saveConfig")
 
-        progress.start("updateIP")
-        progress.pause()
-        const updateIP = await promptUserWithTimeout(
-            "\nWould you like to update the local IP in the configuration? (y/N) [Default: N, Timeout: 10s]: ",
-            10000,
-            "n"
-        )
-        progress.resume()
-
-        if (updateIP.toLowerCase() === "y") {
-            progress.pause()
-            const localIP = getLocalIPAddress()
-            updateConfigWithLocalIP(configPath, localIP)
-            progress.resume()
-            progress.log(`Configuration updated with local IP: ${localIP}`, "success")
-        } else {
-            progress.log("Skipping IP update", "info")
-        }
-        progress.complete("updateIP")
-
-        progress.start("startServer")
-        const configFile = fs.readFileSync(configPath, "utf8")
-        const configObj = JSON.parse(configFile)
-        const port = configObj.WEBVIEW_CONFIG?.port || 3005
-        const running = await isServerRunning(port)
-
-        if (running) {
-            progress.log(`Dev server already running on port ${port}`, "info")
-        } else {
-            progress.pause()
-            const startServer = await promptUserWithTimeout(
-                "\nWould you like to start the dev server? (y/N) [Default: N, Timeout: 10s]: ",
-                10000,
-                "n"
-            )
-            progress.resume()
-
-            if (startServer.toLowerCase() === "y") {
-                startServerBackground()
-                progress.log("Dev server started in background", "success")
-            } else {
-                progress.log("Skipping server startup", "info")
-                progress.log("To serve pages locally, start the dev server manually.", "info")
-            }
-        }
-        progress.complete("startServer")
+        await handlePostSetupSteps("android", configPath, progress, {
+            getLocalIPAddress,
+            updateConfigWithLocalIP,
+            isServerRunning,
+            startServerBackground,
+        })
 
         progress.printTreeContent("Configuration Explanation", [
             "WEBVIEW_CONFIG: Main configuration object for the WebView setup",
