+++ src/native/setupEmulatorIos.js
@@ -1,6 +1,6 @@
 import { execSync } from "child_process"
 import fs from "fs"
-import { runCommand, promptUser, validateAndCompleteConfig, promptUserWithTimeout } from "./utils.js"
+import { runCommand, promptUser, validateAndCompleteConfig, initializeConfig, saveConfig, handlePostSetupSteps } from "./utils.js"
 import TerminalProgress from "./TerminalProgress.js"
 const { getLocalIPAddress, updateConfigWithLocalIP, isServerRunning, startServerBackground } = require("./setupServer.js")
 
@@ -27,46 +27,6 @@ const progressPaddingConfig = {
 }
 
 const progress = new TerminalProgress(steps, "Catalyst Universal Ios Setup", progressPaddingConfig)
-async function initializeConfig() {
-    const configFile = fs.readFileSync(configPath, "utf8")
-    const config = JSON.parse(configFile)
-    const { WEBVIEW_CONFIG } = config
-
-    if (!WEBVIEW_CONFIG || Object.keys(WEBVIEW_CONFIG).length === 0) {
-        progress.log("WebView Config missing in " + configPath, "error")
-        process.exit(1)
-    }
-
-    if (!WEBVIEW_CONFIG.ios) {
-        WEBVIEW_CONFIG.ios = {}
-    }
-
-    return { WEBVIEW_CONFIG }
-}
-
-async function saveConfig(newConfig) {
-    try {
-        // Read the existing config
-        const existingConfigFile = fs.readFileSync(configPath, "utf8")
-        const existingConfig = JSON.parse(existingConfigFile)
-
-        // Merge the new WEBVIEW_CONFIG with existing config
-        const updatedConfig = {
-            ...existingConfig, // Preserve all existing keys
-            WEBVIEW_CONFIG: {
-                ...existingConfig.WEBVIEW_CONFIG,
-                ios: newConfig.WEBVIEW_CONFIG.ios,
-            }, // Update only WEBVIEW_CONFIG
-        }
-
-        // Save the merged config
-        fs.writeFileSync(configPath, JSON.stringify(updatedConfig, null, 2))
-        progress.log("Configuration saved successfully", "success")
-    } catch (error) {
-        progress.log("Failed to save configuration: " + error.message, "error")
-        process.exit(1)
-    }
-}
 
 async function setupIOSEnvironment() {
     try {
@@ -78,7 +38,7 @@ async function setupIOSEnvironment() {
         progress.complete("platform")
 
         progress.start("config")
-        const { WEBVIEW_CONFIG } = await initializeConfig()
+        await initializeConfig("ios", configPath, progress)
         progress.complete("config")
 
         progress.start("simulator")
@@ -91,54 +51,15 @@ async function setupIOSEnvironment() {
 
         progress.start("saveConfig")
         const config = await validateAndCompleteConfig("ios", configPath)
+        await saveConfig("ios", configPath, { WEBVIEW_CONFIG: config }, progress)
         progress.complete("saveConfig")
 
-        progress.start("updateIP")
-        progress.pause()
-        const updateIP = await promptUserWithTimeout(
-            "\nWould you like to update the local IP in the configuration? (y/N) [Default: N, Timeout: 10s]: ",
-            10000,
-            "n"
-        )
-        progress.resume()
-
-        if (updateIP.toLowerCase() === "y") {
-            progress.pause()
-            const localIP = getLocalIPAddress()
-            updateConfigWithLocalIP(configPath, localIP)
-            progress.resume()
-            progress.log(`Configuration updated with local IP: ${localIP}`, "success")
-        } else {
-            progress.log("Skipping IP update", "info")
-        }
-        progress.complete("updateIP")
-
-        progress.start("startServer")
-        const configFile = fs.readFileSync(configPath, "utf8")
-        const configObj = JSON.parse(configFile)
-        const port = configObj.WEBVIEW_CONFIG?.port || 3005
-        const running = await isServerRunning(port)
-
-        if (running) {
-            progress.log(`Dev server already running on port ${port}`, "info")
-        } else {
-            progress.pause()
-            const startServer = await promptUserWithTimeout(
-                "\nWould you like to start the dev server? (y/N) [Default: N, Timeout: 10s]: ",
-                10000,
-                "n"
-            )
-            progress.resume()
-
-            if (startServer.toLowerCase() === "y") {
-                startServerBackground()
-                progress.log("Dev server started in background", "success")
-            } else {
-                progress.log("Skipping server startup", "info")
-                progress.log("To serve pages locally, start the dev server manually.", "info")
-            }
-        }
-        progress.complete("startServer")
+        await handlePostSetupSteps("ios", configPath, progress, {
+            getLocalIPAddress,
+            updateConfigWithLocalIP,
+            isServerRunning,
+            startServerBackground,
+        })
 
         progress.printTreeContent("Configuration Explanation", [
             "WEBVIEW_CONFIG: Main configuration object for the WebView setup",
